Init extension test, migrate to manifest v3 from v2
