+++ front-end/public/background-wrapper.js
@@ -1,0 +1,6 @@
+/* eslint-disable no-undef */
+try {
+  self.importScripts("background.js");
+} catch (e) {
+  console.error(e);
+}
