+++ front-end/public/manifest.json
@@ -1,10 +1,9 @@
 {
-  "short_name": "React App",
   "name": "Text Savvy",
   "description": "Base Level Extension",
   "version": "0.0.1",
-  "manifest_version": 2,
-  "browser_action": {
+  "manifest_version": 3,
+  "action": {
     "default_popup": "popup.html"
   },
   "icons": {
@@ -12,18 +11,20 @@
     "48": "logo.png",
     "128": "logo.png"
   },
+
   "background": {
-    "scripts": [
-      "eventPage.js"
-    ],
-    "persistent": false
+    "service_worker": "background.js",
+    "type": "module"
   },
+
   "permissions": [
     "contextMenus",
     "notifications",
-    "https://localhost:5000/*",
-    "https://localhost:3000/*",
     "tabs"
+    
   ],
-  "content_security_policy": "script-src 'self' 'unsafe-eval'; object-src 'self'"
+  
+  "content_security_policy": {
+    "extension_pages": "script-src 'self'; object-src 'self'"
+  }
 }
